Make actions become actual components.
While debugging the problem with removeItem action on testing template
it came up the convenience of making actions to be actual components of
type "action".

Consequently I implemented a no-op "action" component that inherits all
functionality of base "component" class.

Nevertheless, and despite it helped to diagnose it, the actual problem
was just that removeItem action is inside a list which is a not yet
implemented component (list class is just an empty class waiting for
implementation...).