Singleton component.
Added new component singleton consisting in a form limited to single
data field.

This is useful because its future export implementation will just export
the value of that field (which may be -or not- a scalar) instead of an
object with a single property so that it is suitable for lists of
scalars.