PLAYGROUND: Added employees footer for better UX.
Specially to make a better example of new autoscroll property on
actions...

Counter-intuitively it is more useful on bottom positioned buttons
because the addition of new ittem may move the button down outside of
viewport.

In case of top positioned buttons it is even noisy (I've just removed it
for this reason) because, if you only need to add single item it's
better to use bottom buton but, if you want to insert more in bulk
autoscrolling to new items every time would become so cumbersome.

For bottom buttons this is not a problem because as soon as upper
content grows, last inserted item (and hence bottom toolbar next to it)
will go to exact same position every time.