Added support for decorators.
NOTE: In preliminary tests (not included in this commit) the simple use
of a decorator (forcing it to be included in bundle) increases bundle
size in about 10K (ES Module, for example, goes from 12 to 22k).

I noticed it escaped to minification (pending to investigate how to
avoid this...) but, anyway, looks like it would be an important cost.

On the other hand, benefits I expect from it would be huge and, after
pending core components implementation would not represent so big
percentage over the total.

...and it also will help to avoid code repetition (form and list folding
support will be moved to decorators to avoid code repetition...).