Global and Local events.
  👉 Refactorizend events implementation to decorator instead of base
     class.
     ➡️  As a base class its constructor were getting executed before
        derived classes ones so it were unaware of most SmartForm metadata
        such as ancestry.

  👉 Mantain "two" event handler queues: one listening only to "local"
     events (those triggered from current component) and another stored
     in root component and shared across all components.

  👉 Redesigned old .on() method becoming three new ones:
     ➡️  .onLocal(): To listen locally generated events only.
     ➡️  .onAll(): To listen all events (of specified type).
     ➡️  .on(): Handy alias pointing to onAll() in root component and to
        .onLocal() in the rest.

This way we keep the ability to listen given component's events while
tackling the more useful ability to handle all events of a given type
with single event handler.

That is: If you want to do something when new item is added or removed
from a list, you would probably want to do the same for all lists and
not for single one only.