Starting new branch "fullAsync".
Regarding the decision of keeping components' export() method sync,
lately I have come to realize that it could eventually export
inconsistent data if, for instance, it is called while an import is in
progress.

Specially if there are "addItem" event handlers delaying list items
insertion, for example, awaiting for a rendering css transition.

Since it doesn't look to be necessarily a big issue, I opted to rename
original export() methods to exportSync() and then provide an async
version which just calls the synchronous one but it also shares the
mutex we already put for addItem() and removeItem() actions.

But the same issue applies to methods such as isEmpty() even, having it
just returns a boolean, it doesn't look so harmful if it's value changes
sooner or later.

However, the change hadn't been so strightforward and I ran in a few
async issues with import() functionality (YES: I said import(): export()
works perfectly but now, when importing data for lists, looks like, even
we add the necessary items to it before importing, it doesn't fill the
values until we repeat the import -looks like some issue for not
properly awaiting new components renderisation).

So, at the end, I started to belive that our best bet would be to move
to fully async operation.

But, having this doesn't work propperly yet, I decided to leave it apart
in a separate branch until I could dig deeper into it and make it
consistent.

AMEND:

Doing a few regression test I realized this bug is present from first
addItem event handlers implementation on following commit:

e25c8df5bee15a6fd16a040b24c9824c774cc9a3

...it's just I didn't noticed it.

(We need tests, We need tests, We need tests, We need tests, We need
 tests, We need tests, We need tests, We need tests, We need tests, We
 need tests...)

I'll keep this branch anyway for a while in order to look closer on it
after fixing the bug and before merging again to master.

Nevertheless I still consider it would be better to bet for fully async
operation.