Lists' count action improvement and bugfix.
Starting to tinker with automated tests (not in this commit) noticed it
would be useful for lists' count() action to actually return items count
(not joking: Originally it was there just to allow 'action="count"'
components to exist in order to be injected by lists every time items
are added or removed).

Then I realised that that count actions updating process got
acciedentally removed at some point. So I restored it making counters to
work properly now.