👍 Minor UX fix in playground template.
Make cancel button to reset form even if evaluates to empty (because if
we add items to a list but don't fill any field it's still evaluated as
empty but we may want to tidy all up).

But it allow it without confirm unless form does not evalutate as empty.