Minor changes.
Also moved "company" subform fields outside and removed it in playground
template.

👉 Here it is why:

I grouped then in the past just to test subform grouping and left that
way with the idea of developing kind of "fieldset" component which which
would "flatten" its fields togegher with its parent form on
import/export processes. All of this just for the sake of being able to
fold/unfold arbitrary portions of any form whithout grouping that data
in a new object level.

Now I started the implementation of that idea but meanwhile I realized
it is, in fact, a bad idea due to the fact it would break the
correlation between component paths and the position of their data in
exported JSON.

And, in fact, folding and unfolding arbitrary portions of a form is a
layout issue and can be done with a few CSS and JS if necessary. So it's
definitely out of the SmartForm's scope.