✍️  Whole project rename to SmarkForm.
Even I carefully checked that "SmartForm" name weren't taken in NPM
registry, now, when I finally decided to publish it NPM claims that
there exists a (codeless 😡) package called "smart-form" which is too
similar to "smartform" and does not allow me to publish it with this
name.

So finally I decided to rename it to "SmarkForm".

  👉 Replaced almost everywhere the string "SmartForm" by "SmarkForm"
     (except demo video which is so crappy so I must recreate later
     anyway).

  👉 Renamed "data-smart" special property to "data-smark" (and replaced
     it everywhere with its new name).