🔧🔧 "Action" ➡️  "Trigger" type (disambiguation).
Renamed the previously called "action" type to "trigger".

I have been deferring this change due to the expected underlying
complexity. However, in the end, it turned out to be much easier than
anticipated, as I realized that only the type name (formerly known as
'action') needed to be renamed, since the property, also named 'action',
can (and should) keep its original name having, in fact, it specifies
the ACTIONS to be triggered on other components.

Keeping the name of the "action" property not only drastically reduced
the amount of code to be reviewed but rather, being the declaration of
the "action" (now "trigger") type optional (and even discouraged) in the
presence of the "action" property all existing examples continue working
with no change.

So now, the only thing left to do is to review and update the
documentation, finally eliminating the confusion between the "action"
and "action component type" (now "trigger component type" - or just
"trigger") expressions.