📖 🔧 Workaround to 'dev' script.
After a (necessary) updata I faced that the 'npm run dev' script no
longer worked.

It was because now jekyll tries to start the dev server with ssl.

I managed to do so, but then the --livereload started to fail.

...fially I discovered it has dependencies that cannot be run without
ssl.

👉 For this reason I had to degrade the dev experience from --livereload
to --watch which still rebuilds after every change, but requires manual
reload.